Fix dtype mismatch in subgoal head and add validation script
- Convert manager_repr to float32 before passing to subgoal head
- Add comprehensive validation script for feudal loss integration
- All validation tests pass

+++ HRM/models/hrm/hrm_act_v1.py
@@ -430,7 +430,7 @@ class HierarchicalReasoningModel_ACTV1(nn.Module):
         subgoal_output = None
         if self.subgoal_head is not None:
             manager_hidden = extras["manager_hidden"]
-            manager_repr = manager_hidden[:, 0]
+            manager_repr = manager_hidden[:, 0].to(torch.float32)
             new_subgoal_state, subgoal_output = self.subgoal_head(
                 manager_repr, subgoal_state
             )
